Add support for validating dashboard load configuration and split models into separate report and dashboard modules. (#13)
