+++ src/schemas/loadDashboard.json
@@ -1,0 +1,26 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "type": "object",
+  "properties": {
+    "accessToken": {
+      "type": "string",
+      "messages": {
+        "type": "accessToken must be a string",
+        "required": "accessToken is required"
+      },
+      "invalidMessage": "accessToken property is invalid"
+    },
+    "id": {
+      "type": "string",
+      "messages": {
+        "type": "id must be a string",
+        "required": "id is required"
+      },
+      "invalidMessage": "id property is invalid"
+    }
+  },
+  "required": [
+    "accessToken",
+    "id"
+  ]
+}
