+++ test/models.spec.ts
@@ -1,23 +1,121 @@
 import * as models from '../src/models';
 
-describe('Unit | Models', function () {
+describe('Unit | models.dashboard', function () {
+
+    function testForExpectedMessage(errors: models.IError[], message: string) {
+      expect(errors).toBeDefined();
+      let flag = false;
+      errors.forEach(error => {
+        if (error.message === message) {
+            flag = true;
+        }
+      });
+	  
+	  expect(flag).toBe(true);
+    }
+  
+  describe('validateLoad', function () {
+
+    const accessTokenRequiredMessage = models.dashboard.loadSchema.properties.accessToken.messages.required;
+    const accessTokenInvalidTypeMessage = models.dashboard.loadSchema.properties.accessToken.invalidMessage;
+    const idRequiredMessage = models.dashboard.loadSchema.properties.id.messages.required;
+    const idInvalidTypeMessage = models.dashboard.loadSchema.properties.id.invalidMessage;
+
+    it(`should return errors with one containing message '${accessTokenRequiredMessage}' if accessToken is not defined`, function () {
+        // Arrange
+        const testData = {
+            load: {
+            }
+        };
+
+        // Act
+        const errors = models.dashboard.validateLoad(testData.load);
+
+        // Assert
+        testForExpectedMessage(errors, accessTokenRequiredMessage);
+    });
+
+    it(`should return errors with one containing message '${accessTokenInvalidTypeMessage}' if accessToken is not a string`, function () {
+        // Arrange
+        const testData = {
+            load: {
+                accessToken: 1
+            }
+        };
+
+        // Act
+        const errors = models.dashboard.validateLoad(testData.load);
+
+        // Assert
+        testForExpectedMessage(errors, accessTokenInvalidTypeMessage);
+    });
+
+    it(`should return errors with one containing message '${idRequiredMessage}' if id is not defined`, function () {
+        // Arrange
+        const testData = {
+            load: {
+                accessToken: '123'
+            }
+        };
+
+        // Act
+        const errors = models.dashboard.validateLoad(testData.load);
+
+        // Assert
+        testForExpectedMessage(errors, idRequiredMessage);
+    });
+
+    it(`should return errors with one containing message '${idInvalidTypeMessage}' if id is not a string`, function () {
+        // Arrange
+        const testData = {
+           load: {
+               id: 1
+           }
+        };
+
+        // Act
+        const errors = models.dashboard.validateLoad(testData.load);
+
+        // Assert
+        testForExpectedMessage(errors, idInvalidTypeMessage);
+    });
+
+    it(`should return undefined if id and accessToken are provided`, function () {
+        // Arrange
+        const testData = {
+            load: {
+               id: 'fakeId',
+               accessToken: 'fakeAccessToken'
+            }
+        };
+
+        // Act
+        const errors = models.dashboard.validateLoad(testData.load);
+
+        // Assert
+        expect(errors).toBeUndefined();
+     });
+  });
+});
+
+describe('Unit | models.report', function () {
   function testForExpectedMessage(errors: models.IError[], message: string) {
     expect(errors).toBeDefined();
     errors
       .forEach(error => {
         if (error.message === message) {
-          expect(true).toBe(true);
+            expect(true).toBe(true);
         }
       });
   }
-
+  
   describe('validateLoad', function () {
-    const accessTokenRequiredMessage = models.loadSchema.properties.accessToken.messages.required;
-    const accessTokenInvalidTypeMessage = models.loadSchema.properties.accessToken.messages.type;
-    const idRequiredMessage = models.loadSchema.properties.id.messages.required;
-    const idInvalidTypeMessage = models.loadSchema.properties.id.messages.type;
-    const filtersInvalidMessage = models.loadSchema.properties.filters.invalidMessage;
-    const pageNameInvalidTypeMessage = models.loadSchema.properties.pageName.messages.type;
+    const accessTokenRequiredMessage = models.report.loadSchema.properties.accessToken.messages.required;
+    const accessTokenInvalidTypeMessage = models.report.loadSchema.properties.accessToken.messages.type;
+    const idRequiredMessage = models.report.loadSchema.properties.id.messages.required;
+    const idInvalidTypeMessage = models.report.loadSchema.properties.id.messages.type;
+    const filtersInvalidMessage = models.report.loadSchema.properties.filters.invalidMessage;
+    const pageNameInvalidTypeMessage = models.report.loadSchema.properties.pageName.messages.type;
 
     it(`should return errors with one containing message '${accessTokenRequiredMessage}' if accessToken is not defined`, function () {
       // Arrange
@@ -27,7 +125,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateLoad(testData.load);
+      const errors = models.report.validateLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, accessTokenRequiredMessage);
@@ -42,7 +140,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateLoad(testData.load);
+      const errors = models.report.validateLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, accessTokenInvalidTypeMessage);
@@ -56,7 +154,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateLoad(testData.load);
+      const errors = models.report.validateLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, idRequiredMessage);
@@ -70,7 +168,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateLoad(testData.load);
+      const errors = models.report.validateLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, idRequiredMessage);
@@ -86,7 +184,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateLoad(testData.load);
+      const errors = models.report.validateLoad(testData.load);
 
       // Assert
       expect(errors).toBeUndefined();
@@ -103,7 +201,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateLoad(testData.load);
+      const errors = models.report.validateLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, filtersInvalidMessage);
@@ -120,7 +218,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateLoad(testData.load);
+      const errors = models.report.validateLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, pageNameInvalidTypeMessage);
@@ -157,9 +255,9 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateFilter(malformedFilter);
-      const errors1 = models.validateFilter(malformedFilter1);
-      const errors2 = models.validateFilter(malformedFilter2);
+      const errors = models.report.validateFilter(malformedFilter);
+      const errors1 = models.report.validateFilter(malformedFilter1);
+      const errors2 = models.report.validateFilter(malformedFilter2);
 
       // Assert
       expect(errors).toBeDefined();
@@ -169,7 +267,7 @@ describe('Unit | Models', function () {
 
     it("should return undefined if object is valid basic filter schema", function () {
       // Arrange
-      const expectedFilter: models.IBasicFilter = {
+      const expectedFilter: models.report.IBasicFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -184,18 +282,18 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const filter = new models.BasicFilter(
+      const filter = new models.report.BasicFilter(
         expectedFilter.target,
         expectedFilter.operator,
         expectedFilter.values);
 
       // Assert
-      expect(models.validateFilter(filter.toJSON())).toBeUndefined();
+      expect(models.report.validateFilter(filter.toJSON())).toBeUndefined();
     });
 
     it("should return undefined if object is valid advanced filter schema", function () {
       // Arrange
-      const expectedFilter: models.IAdvancedFilter = {
+      const expectedFilter: models.report.IAdvancedFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -218,19 +316,19 @@ describe('Unit | Models', function () {
         ]
       };
 
-      const filter = new models.AdvancedFilter(
+      const filter = new models.report.AdvancedFilter(
         expectedFilter.target,
         expectedFilter.logicalOperator,
         ...expectedFilter.conditions.slice(0, 2));
 
-      const filter2 = new models.AdvancedFilter(
+      const filter2 = new models.report.AdvancedFilter(
         expectedFilter.target,
         expectedFilter.logicalOperator,
         ...expectedFilter.conditions.slice(1, 3));
 
       // Act
-      const errors = models.validateFilter(filter.toJSON());
-      const errors2 = models.validateFilter(filter2.toJSON());
+      const errors = models.report.validateFilter(filter.toJSON());
+      const errors2 = models.report.validateFilter(filter2.toJSON());
 
       // Assert
       expect(errors).toBeUndefined();
@@ -239,8 +337,8 @@ describe('Unit | Models', function () {
   });
 
   describe('validateSettings', function () {
-    const filterPaneEnabledInvalidTypeMessage = models.settingsSchema.properties.filterPaneEnabled.messages.type;
-    const navContentPaneEnabledInvalidTypeMessage = models.settingsSchema.properties.navContentPaneEnabled.messages.type;
+    const filterPaneEnabledInvalidTypeMessage = models.report.settingsSchema.properties.filterPaneEnabled.messages.type;
+    const navContentPaneEnabledInvalidTypeMessage = models.report.settingsSchema.properties.navContentPaneEnabled.messages.type;
 
     it(`should return errors with one containing message '${filterPaneEnabledInvalidTypeMessage}' if filterPaneEnabled is not a boolean`, function () {
       // Arrange
@@ -251,7 +349,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateSettings(testData.settings);
+      const errors = models.report.validateSettings(testData.settings);
 
       // Assert
       testForExpectedMessage(errors, filterPaneEnabledInvalidTypeMessage);
@@ -266,7 +364,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateSettings(testData.settings);
+      const errors = models.report.validateSettings(testData.settings);
 
       // Assert
       testForExpectedMessage(errors, navContentPaneEnabledInvalidTypeMessage);
@@ -280,7 +378,7 @@ describe('Unit | Models', function () {
       };
 
       // Act
-      const errors = models.validateSettings(testData.settings);
+      const errors = models.report.validateSettings(testData.settings);
 
       // Assert
       expect(errors).toBeUndefined();
@@ -294,7 +392,7 @@ describe("Unit | Filters", function () {
       // Arrange
 
       // Act
-      const basicFilter = new models.BasicFilter({ table: "t", column: "c" }, "In", 1, 2);
+      const basicFilter = new models.report.BasicFilter({ table: "t", column: "c" }, "In", 1, 2);
 
       // Assert
       expect(basicFilter.values).toEqual([1, 2]);
@@ -305,7 +403,7 @@ describe("Unit | Filters", function () {
       const values = [1, 2];
 
       // Act
-      const basicFilter = new models.BasicFilter({ table: "t", column: "c" }, "In", values);
+      const basicFilter = new models.report.BasicFilter({ table: "t", column: "c" }, "In", values);
 
       // Assert
       expect(basicFilter.values).toEqual(values);
@@ -313,7 +411,7 @@ describe("Unit | Filters", function () {
 
     it("should return valid json format when toJSON is called", function () {
       // Arrange
-      const expectedFilter: models.IBasicFilter = {
+      const expectedFilter: models.report.IBasicFilter = {
         $schema: "http://powerbi.com/product/schema#basic",
         target: {
           table: "a",
@@ -328,7 +426,7 @@ describe("Unit | Filters", function () {
       };
 
       // Act
-      const filter = new models.BasicFilter(
+      const filter = new models.report.BasicFilter(
         expectedFilter.target,
         expectedFilter.operator,
         expectedFilter.values);
@@ -339,7 +437,7 @@ describe("Unit | Filters", function () {
 
     it("can be constructed using either array form or individual arguments", function () {
       // Arrange
-      const expectedFilter: models.IBasicFilter = {
+      const expectedFilter: models.report.IBasicFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -354,8 +452,8 @@ describe("Unit | Filters", function () {
       };
 
       // Act
-      const filter1 = new models.BasicFilter(expectedFilter.target, expectedFilter.operator, expectedFilter.values);
-      const filter2 = new models.BasicFilter(expectedFilter.target, expectedFilter.operator, ...expectedFilter.values);
+      const filter1 = new models.report.BasicFilter(expectedFilter.target, expectedFilter.operator, expectedFilter.values);
+      const filter2 = new models.report.BasicFilter(expectedFilter.target, expectedFilter.operator, ...expectedFilter.values);
 
       // Assert
       expect(filter1.toJSON()).toEqual(filter2.toJSON());
@@ -365,14 +463,14 @@ describe("Unit | Filters", function () {
   describe("AdvancedFilter", function () {
     it("should throw an error if logical operator is not a non-empty string", function () {
       // Arrange
-      const condition: models.IAdvancedFilterCondition = {
+      const condition: models.report.IAdvancedFilterCondition = {
         value: "a",
         operator: "LessThan"
       };
 
       // Act
       const attemptToCreateFilter = () => {
-        return new models.AdvancedFilter({ table: "t", column: "c" }, <any>1, condition);
+        return new models.report.AdvancedFilter({ table: "t", column: "c" }, <any>1, condition);
       };
 
       // Assert
@@ -381,7 +479,7 @@ describe("Unit | Filters", function () {
 
     it("should throw an error if more than two conditions are provided", function () {
       // Arrange
-      const conditions: models.IAdvancedFilterCondition[] = [
+      const conditions: models.report.IAdvancedFilterCondition[] = [
         {
           value: "a",
           operator: "LessThan"
@@ -398,7 +496,7 @@ describe("Unit | Filters", function () {
 
       // Act
       const attemptToCreateFilter = () => {
-        return new models.AdvancedFilter({ table: "Table", column: "c" }, "And", ...conditions);
+        return new models.report.AdvancedFilter({ table: "Table", column: "c" }, "And", ...conditions);
       };
 
       // Assert
@@ -410,7 +508,7 @@ describe("Unit | Filters", function () {
 
       // Act
       const attemptToCreateFilter = () => {
-        return new models.AdvancedFilter({ table: "Table", column: "c" }, "Or", { value: "a", operator: "Contains" });
+        return new models.report.AdvancedFilter({ table: "Table", column: "c" }, "Or", { value: "a", operator: "Contains" });
       };
 
       // Assert
@@ -419,7 +517,7 @@ describe("Unit | Filters", function () {
 
     it("should output the correct json when toJSON is called", function () {
       // Arrange
-      const expectedFilter: models.IAdvancedFilter = {
+      const expectedFilter: models.report.IAdvancedFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -439,7 +537,7 @@ describe("Unit | Filters", function () {
       };
 
       // Act
-      const filter = new models.AdvancedFilter(
+      const filter = new models.report.AdvancedFilter(
         expectedFilter.target,
         expectedFilter.logicalOperator,
         ...expectedFilter.conditions);
@@ -450,7 +548,7 @@ describe("Unit | Filters", function () {
 
     it("can be constructed using either array form or individual arguments", function () {
       // Arrange
-      const expectedFilter: models.IAdvancedFilter = {
+      const expectedFilter: models.report.IAdvancedFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -470,8 +568,8 @@ describe("Unit | Filters", function () {
       };
 
       // Act
-      const filter1 = new models.AdvancedFilter(expectedFilter.target, expectedFilter.logicalOperator, expectedFilter.conditions);
-      const filter2 = new models.AdvancedFilter(expectedFilter.target, expectedFilter.logicalOperator, ...expectedFilter.conditions);
+      const filter1 = new models.report.AdvancedFilter(expectedFilter.target, expectedFilter.logicalOperator, expectedFilter.conditions);
+      const filter2 = new models.report.AdvancedFilter(expectedFilter.target, expectedFilter.logicalOperator, ...expectedFilter.conditions);
 
       // Assert
       expect(filter1.toJSON()).toEqual(filter2.toJSON());
@@ -482,20 +580,20 @@ describe("Unit | Filters", function () {
     it('getFilterType should return type of filter given a filter object', function () {
       // Arrange
       const testData = {
-        basicFilter: new models.BasicFilter({ table: "a", column: "b" }, "In", ["x", "y"]),
-        advancedFilter: new models.AdvancedFilter({ table: "a", column: "b" }, "And",
+        basicFilter: new models.report.BasicFilter({ table: "a", column: "b" }, "In", ["x", "y"]),
+        advancedFilter: new models.report.AdvancedFilter({ table: "a", column: "b" }, "And",
           { operator: "Contains", value: "x" },
           { operator: "Contains", value: "x" }
         ),
-        nonFilter: <models.IFilter>{}
+        nonFilter: <models.report.IFilter>{}
       };
 
       // Act
 
       // Assert
-      expect(models.getFilterType(testData.basicFilter.toJSON())).toBe(models.FilterType.Basic);
-      expect(models.getFilterType(testData.advancedFilter.toJSON())).toBe(models.FilterType.Advanced);
-      expect(models.getFilterType(testData.nonFilter)).toBe(models.FilterType.Unknown);
+      expect(models.report.getFilterType(testData.basicFilter.toJSON())).toBe(models.report.FilterType.Basic);
+      expect(models.report.getFilterType(testData.advancedFilter.toJSON())).toBe(models.report.FilterType.Advanced);
+      expect(models.report.getFilterType(testData.nonFilter)).toBe(models.report.FilterType.Unknown);
     });
   });
 });
